NEW - bug 103932: support applying patches via drag-and-drop https://bugs.eclipse.org/bugs/show_bug.cgi?id=103932

+++ org.eclipse.mylyn.team.ui/src/org/eclipse/mylyn/internal/team/ui/actions/ApplyPatchAction.java
@@ -50,50 +50,55 @@ public class ApplyPatchAction implements IViewActionDelegate {
 			if (object instanceof RepositoryAttachment) {
 				RepositoryAttachment attachment = (RepositoryAttachment) object;
 				if (attachment.getRepository() == null) {
-					MessageDialog.openInformation(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getShell(), 
-							TasksUiPlugin.TITLE_DIALOG, 
-							"Please synchronize this task in order to apply the patch.");
+					MessageDialog.openInformation(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getShell(),
+							TasksUiPlugin.TITLE_DIALOG, "Please synchronize this task in order to apply the patch.");
 				} else {
 					final String contents = TasksUiPlugin.getRepositoryManager().getAttachmentContents(attachment);
-					// TODO: shouldn't need to do this
-					IPath statePath = MylarTeamPlugin.getDefault().getStateLocation();
-					final File tempFile = new File(statePath.toString() + File.separator + "patch.txt");
-					tempFile.deleteOnExit();
-					try {
-						BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile));
-						writer.write(contents);
-//						System.err.println(">>> " + contents);
-						writer.close();
-					} catch (Exception e) {
-						MylarStatusHandler.fail(e, "could not write patch file", true);
-					} 
-					
-					IStorage storage = new IStorage() {
-
-						@SuppressWarnings("deprecation")
-						public InputStream getContents() throws CoreException {
-							return new StringBufferInputStream(contents);
+					if (contents == null) {
+						MessageDialog.openInformation(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getShell(),
+								TasksUiPlugin.TITLE_DIALOG, "Patch could not be retrieved.");
+					} else {
+						// TODO: shouldn't need to do this
+						IPath statePath = MylarTeamPlugin.getDefault().getStateLocation();
+						final File tempFile = new File(statePath.toString() + File.separator + "patch.txt");
+						tempFile.deleteOnExit();
+						try {
+							BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile));
+							writer.write(contents);
+							// System.err.println(">>> " + contents);
+							writer.close();
+						} catch (Exception e) {
+							MylarStatusHandler.fail(e, "could not write patch file", true);
 						}
 
-						public IPath getFullPath() {
-							return new Path(tempFile.getAbsolutePath());
-						}
+						IStorage storage = new IStorage() {
 
-						public String getName() {
-							return null;
-						}
+							@SuppressWarnings("deprecation")
+							public InputStream getContents() throws CoreException {
+								return new StringBufferInputStream(contents);
+							}
 
-						public boolean isReadOnly() {
-							return true;
-						}
+							public IPath getFullPath() {
+								return new Path(tempFile.getAbsolutePath());
+							}
 
-						public Object getAdapter(Class adapter) {
-							return null;
-						}
+							public String getName() {
+								return null;
+							}
 
-					};
-					ApplyPatchOperation op = new ApplyPatchOperation(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage().getActivePart(), storage, null, null);
-					BusyIndicator.showWhile(Display.getDefault(), op);
+							public boolean isReadOnly() {
+								return true;
+							}
+
+							public Object getAdapter(Class adapter) {
+								return null;
+							}
+
+						};
+						ApplyPatchOperation op = new ApplyPatchOperation(PlatformUI.getWorkbench()
+								.getActiveWorkbenchWindow().getActivePage().getActivePart(), storage, null, null);
+						BusyIndicator.showWhile(Display.getDefault(), op);
+					}
 				}
 			}
 		}
